pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "navis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eclipse-zenoh>=1.5.1",
    "matplotlib>=3.10.6",
    "msgspec>=0.19.0",
    "numpy>=2.3.3",
    "shibuya>=2025.9.25",
    "sphinx>=8.2.3",
    "sphinx-design>=0.6.1",
    "sphinxcontrib-video>=0.4.1",
]

[project.scripts]
navis-router = "navis.router:main"
navis-visualizer = "navis.visualizer:main"
